+++ webrtc/src/webrtc_c/api/video/i420_buffer.h
@@ -13,6 +13,8 @@ extern "C" {
 // -------------------------
 
 WEBRTC_DECLARE_REFCOUNTED(webrtc_I420Buffer);
+struct webrtc_VideoFrameBuffer_refcounted;
+
 WEBRTC_EXPORT struct webrtc_I420Buffer_refcounted* webrtc_I420Buffer_Create(
     int width,
     int height);
@@ -30,6 +32,7 @@ WEBRTC_EXPORT int webrtc_I420Buffer_StrideU(struct webrtc_I420Buffer* self);
 WEBRTC_EXPORT int webrtc_I420Buffer_StrideV(struct webrtc_I420Buffer* self);
 WEBRTC_EXPORT void webrtc_I420Buffer_ScaleFrom(struct webrtc_I420Buffer* self,
                                                struct webrtc_I420Buffer* src);
+WEBRTC_DECLARE_CAST_REFCOUNTED(webrtc_I420Buffer, webrtc_VideoFrameBuffer);
 
 #if defined(__cplusplus)
 }