+++ webrtc/src/webrtc_c/api/video/video_frame.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 #include "../../common.h"
-#include "i420_buffer.h"
+#include "video_frame_buffer.h"
 
 #if defined(__cplusplus)
 extern "C" {
@@ -20,7 +20,7 @@ WEBRTC_EXPORT extern const int webrtc_VideoRotation_270;
 
 WEBRTC_DECLARE_UNIQUE(webrtc_VideoFrame);
 WEBRTC_EXPORT struct webrtc_VideoFrame_unique* webrtc_VideoFrame_Create(
-    struct webrtc_I420Buffer_refcounted* buffer,
+    struct webrtc_VideoFrameBuffer_refcounted* buffer,
     int rotation,
     int64_t timestamp_us,
     uint32_t timestamp_rtp);
@@ -33,7 +33,7 @@ WEBRTC_EXPORT uint32_t
 webrtc_VideoFrame_timestamp_rtp(const struct webrtc_VideoFrame* self);
 WEBRTC_EXPORT int webrtc_VideoFrame_rotation(
     const struct webrtc_VideoFrame* self);
-WEBRTC_EXPORT struct webrtc_I420Buffer_refcounted*
+WEBRTC_EXPORT struct webrtc_VideoFrameBuffer_refcounted*
 webrtc_VideoFrame_video_frame_buffer(const struct webrtc_VideoFrame* self);
 
 #if defined(__cplusplus)